refactor: extract reportStatus into its own tap module
Move the CLI-native status command (reportStatus and its helpers) out of
the tapModule file into cli/lib/tap/status.ts. The shared exec-envelope
validator moves to tap-session.ts so both tap.ts and status.ts import it
one-directionally, avoiding a cycle.
